COM-3684: move files
