+++ src/app/Courses/ActivityCardContainer/cardTemplates/ActivityStartCard/styles.ts
