+++ src/app/Courses/QuestionnaireCardContainer/cardTemplates/QuestionnaireStartCard/_layout.tsx
@@ -1,8 +1,8 @@
 import { Stack } from 'expo-router';
 
-const StartCardLayout = () => (
+const QuestionnaireStartCardLayout = () => (
   <Stack screenOptions={{ headerShown: false }}>
     <Stack.Screen name='index'/>
   </Stack>
 );
-export default StartCardLayout;
+export default QuestionnaireStartCardLayout;
