+++ src/components/cardTemplates/FillTheGapCard/index.tsx
