+++ src/components/cardTemplates/FillTheGapCard/styles.ts
