+++ src/components/cardTemplates/FlashCard/index.tsx
