+++ src/components/cardTemplates/FlashCard/styles.ts
