+++ src/components/cardTemplates/MultipleChoiceQuestionCard/styles.ts
