+++ src/components/cardTemplates/OpenQuestionCard/index.tsx
