+++ src/components/cardTemplates/OpenQuestionCard/styles.ts
