+++ src/components/cardTemplates/OrderTheSequenceCard/index.tsx
