+++ src/components/cardTemplates/OrderTheSequenceCard/styles.ts
