+++ src/components/cardTemplates/QuestionAnswerCard/index.tsx
