+++ src/components/cardTemplates/QuestionAnswerCard/styles.ts
