+++ src/components/cardTemplates/SingleChoiceQuestionCard/index.tsx
