+++ src/components/cardTemplates/SingleChoiceQuestionCard/styles.ts
