+++ src/components/cardTemplates/SurveyCard/styles.ts
