+++ src/components/cardTemplates/TextMediaCard/index.tsx
