+++ src/components/cardTemplates/TextMediaCard/styles.ts
