+++ src/components/cardTemplates/TitleTextCard/styles.ts
