+++ src/components/cardTemplates/TitleTextMediaCard/index.tsx
