+++ src/components/cardTemplates/TitleTextMediaCard/styles.ts
