+++ src/components/cardTemplates/Transition/styles.ts
